TWG DH KR --> Extract metadata from KR files, store & determine table linkage
Added support for table matching based on actual field name (last 4
chars) followed by verfication of Field Desc. Also determined the final
no of tables which can be joint intially, based on the FKs present in
the table.

Not added support to enable inference i.e. A has FKs for B. B has FKs
for C. Hence, A can be connected to C. Working on the same.